requirements.txt
numpy==1.19.5
pandas==1.1.3
seaborn==0.11.0
scikit-learn==0.23.2
scipy==1.4.1
statsmodels==0.11.1
plotly>=4
tensorflow
librosa
pydub
matplotlib
IPython
#cv2
plotly